update.sh: swap mem and mmio regions (mem starts at 0x8000_0000)
The total size of RAM can be expected to vary significantly, and
often significantly exceed the size needed for MMIO allocations.

As such, it makes sense to "confine" MMIO to addresses below the
0x8000_0000 boundary, and allow RAM to grow without impediment
from that address upward.

This patch swaps the current MMIO and MEM allocations for LiteX
specific Rocket variants, routing memory accesses in the range
from [0x1000_0000..0x7fff_ffff] to the (uncached) MMIO AXI port,
and from [0x8000_0000..0xffff_ffff] to the (cached) MEM AXI, from
where it can be extended further in the future, on systems with
RAM in excess of 2GB.

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>